Move Auth Updater Side Effect to a Global Wrapper
